cool table css

+++ db/migrate/20131211220945_create_transactions.rb
@@ -3,7 +3,7 @@ class CreateTransactions < ActiveRecord::Migration
     create_table :transactions do |t|
       t.integer  :user_id, null: false
       t.string   :description, null: false
-      t.decimal  :amount, null: false, scale: 2
+      t.float  :amount, null: false, scale: 2
       t.string   :category
       t.string   :privacy
       t.string   :date, null: false
